requirements.txt
asgiref==3.8.1
contourpy==1.3.0
cvxopt==1.3.2
cycler==0.12.1
deprecation==2.1.0
Django==5.1.1
fonttools==4.54.1
graphviz==0.20.3
intervaltree==3.1.0
jmetalpy==1.7.0
kiwisolver==1.4.7
lxml==5.3.0
matplotlib==3.9.2
mockito==1.5.1
networkx==3.3
numpy==2.1.1
packaging==24.1
pandas==2.2.3
patsy==0.5.6
pillow==10.4.0
plotly==5.24.1
pm4py==2.7.11.13
pydotplus==2.0.2
PyHamcrest==2.1.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
pytz==2024.2
scipy==1.14.1
six==1.16.0
sortedcontainers==2.4.0
sqlparse==0.5.1
statsmodels==0.14.3
tenacity==9.0.0
tqdm==4.66.5
tzdata==2024.2
